examples/runner: refactor for clarity
- Updated session service initialization in main.go to use a more structured approach with error handling for Redis and inmemory options.
- Enhanced README.md to clarify prerequisites and command line arguments, including a correction to the Go version requirement.
- Improved output formatting in the main application for better readability and consistency.
