+++ examples/runner/tools.go
@@ -1,0 +1,106 @@
+package main
+
+import (
+	"context"
+	"strings"
+	"time"
+)
+
+// calculate performs basic mathematical operations.
+func (c *multiTurnChat) calculate(ctx context.Context, args calculatorArgs) (calculatorResult, error) {
+	var result float64
+
+	switch strings.ToLower(args.Operation) {
+	case "add", "+":
+		result = args.A + args.B
+	case "subtract", "-":
+		result = args.A - args.B
+	case "multiply", "*":
+		result = args.A * args.B
+	case "divide", "/":
+		if args.B != 0 {
+			result = args.A / args.B
+		} else {
+			result = 0 // Handle division by zero.
+		}
+	default:
+		result = 0
+	}
+
+	return calculatorResult{
+		Operation: args.Operation,
+		A:         args.A,
+		B:         args.B,
+		Result:    result,
+	}, nil
+}
+
+// getCurrentTime returns current time information.
+func (c *multiTurnChat) getCurrentTime(_ context.Context, args timeArgs) (timeResult, error) {
+	now := time.Now()
+	var t time.Time
+	timezone := args.Timezone
+
+	// Handle timezone conversion.
+	switch strings.ToUpper(args.Timezone) {
+	case "UTC":
+		t = now.UTC()
+	case "EST", "EASTERN":
+		t = now.Add(-5 * time.Hour) // Simplified EST.
+	case "PST", "PACIFIC":
+		t = now.Add(-8 * time.Hour) // Simplified PST.
+	case "CST", "CENTRAL":
+		t = now.Add(-6 * time.Hour) // Simplified CST.
+	case "":
+		t = now
+		timezone = "Local"
+	default:
+		t = now.UTC()
+		timezone = "UTC"
+	}
+
+	return timeResult{
+		Timezone: timezone,
+		Time:     t.Format("15:04:05"),
+		Date:     t.Format("2006-01-02"),
+		Weekday:  t.Weekday().String(),
+	}, nil
+}
+
+// calculatorArgs represents arguments for the calculator tool.
+type calculatorArgs struct {
+	Operation string  `json:"operation" jsonschema:"description=The operation: add, subtract, multiply, divide,enum=add,enum=subtract,enum=multiply,enum=divide"`
+	A         float64 `json:"a" jsonschema:"description=First number"`
+	B         float64 `json:"b" jsonschema:"description=Second number"`
+}
+
+// calculatorResult represents the result of a calculation.
+type calculatorResult struct {
+	Operation string  `json:"operation"`
+	A         float64 `json:"a"`
+	B         float64 `json:"b"`
+	Result    float64 `json:"result"`
+}
+
+// timeArgs represents arguments for the time tool.
+type timeArgs struct {
+	Timezone string `json:"timezone" jsonschema:"description=Timezone (UTC, EST, PST, CST) or leave empty for local"`
+}
+
+// timeResult represents the current time information.
+type timeResult struct {
+	Timezone string `json:"timezone"`
+	Time     string `json:"time"`
+	Date     string `json:"date"`
+	Weekday  string `json:"weekday"`
+}
+
+// Helper functions for creating pointers to primitive types.
+
+func intPtr(i int) *int {
+	return &i
+}
+
+func floatPtr(f float64) *float64 {
+	return &f
+}
